+++ Application/DTOs/TestDTO/Commands/CreateTestCommand.cs
@@ -19,7 +19,7 @@ namespace Application.DTOs
         public int? LessonDefinitionId { get; set; }
         public int? SubLevelId { get; set; }
         public int? LevelId { get; set; }
-        public int TotalPoint { get; set; }
+        public double TotalPoint { get; set; }
 
         public class CreateTestCommandHandler : IRequestHandler<CreateTestCommand, Response<int>>
         {
@@ -52,7 +52,7 @@ namespace Application.DTOs
                 test.PlacementStartDate = null;
                 
                 test = await _TestRepository.AddAsync(test);
-                int total = 0;
+                double total = 0;
                 foreach (var item in command.Questions)
                 {
                     item.TestId = test.Id;
